% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_coord_fn.R
\name{apply_coord_fn_}
\alias{apply_coord_fn_}
\title{Apply coordinate function (internal wrapper)}
\usage{
apply_coord_fn_(..., cols, coord_fn, fn_name, coordinate_name)
}
\arguments{
\item{...}{\code{Numeric vectors} or a single \code{data.frame}.}

\item{cols}{Names of columns to use when \code{`...`} is a single \code{data.frame}.}

\item{coord_fn}{Function that takes each \code{vector}/\code{column} as separate args via \code{...} and
returns a \code{vector} with one value per input \code{vector}/\code{column}.}

\item{fn_name}{Name of applied function for messages.}

\item{coordinate_name}{Name of coordinates (e.g. \code{"centroid"}) for messages.}
}
\description{
Apply coordinate function (internal wrapper)
}
\keyword{internal}
