% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{by_distance_rearranger}
\alias{by_distance_rearranger}
\title{Wrapper for running closest to / furthest from rearrange methods}
\usage{
by_distance_rearranger(
  data,
  cols,
  origin = NULL,
  origin_fn = NULL,
  shuffle_ties = FALSE,
  decreasing = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{vector}.}

\item{cols}{Column(s) to create sorting factor by. When \code{NULL} and \code{`data`} is a \code{data.frame},
the row numbers are used.}

\item{origin}{Coordinates of the origin to calculate distances to.
Must be either a single constant to use in all dimensions
or a \code{vector} with one constant per dimension.

\strong{N.B.} Ignored when \code{`origin_fn`} is not \code{NULL}.}

\item{origin_fn}{Function for finding the origin coordinates to calculate distances to.
Each column will be passed as a \code{vector} in the order of \code{`cols`}.
It should return a \code{vector} with one constant per dimension.

Can be created with \code{\link[rearrr:create_origin_fn]{create_origin_fn()}} if you want to apply
the same function to each dimension.

E.g. the \code{\link[rearrr:centroid]{centroid()}} function, which is created with:

\code{create_origin_fn(mean)}

Which returns the following function:

\code{function(...)\{}

\verb{  }\code{list(...) \%>\%}

\verb{    }\code{purrr::map(mean) \%>\%}

\verb{    }\code{unlist(recursive = TRUE,}

\verb{           }\code{use.names = FALSE)}

\code{\}}}

\item{shuffle_ties}{Whether to shuffle elements with the same distance to the origin. (Logical)}

\item{decreasing}{Whether to order by decreasing distances to the origin. (Logical)}
}
\value{
The sorted \code{data.frame} (\code{tibble}) / \code{vector}.
}
\description{
Wrapper for running closest to / furthest from rearrange methods
}
\keyword{internal}
