% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Find the coordinates for the centroid}
\usage{
centroid(..., na.rm = FALSE)
}
\arguments{
\item{...}{Numeric vectors.}

\item{na.rm}{Whether to ignore missing values when calculating means. (Logical)}
}
\value{
\code{vector} with the means of each supplied vector.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Calculates the mean for each of the passed vectors.
}
\examples{
\donttest{
# Attach packages
library(rearrr)

# Set seed
set.seed(1)

# Create three vectors
x <- runif(10)
y <- runif(10)
z <- runif(10)

# Find centroid coordinates
# Aka. the means of each vector
centroid(x, y, z)
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
