% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_origin_fn.R
\name{create_origin_fn}
\alias{create_origin_fn}
\title{Create origin_fn function}
\usage{
create_origin_fn(fn)
}
\arguments{
\item{fn}{Function to apply to each dimension. Should return a numeric scalar.}
}
\value{
Function with the dots (\code{...}) argument that applies the \code{`fn`} function to
each element in \code{...} (usually one vector per dimension). More specifically:

\code{function(...)\{}

\verb{  }\code{list(...) \%>\%}

\verb{    }\code{purrr::map(fn) \%>\%}

\verb{    }\code{unlist(recursive = TRUE,}

\verb{           }\code{use.names = FALSE)}

\code{\}}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Creates a function that applies a supplied function to all input vectors.
}
\examples{
\donttest{
# Attach packages
library(rearrr)

# Set seed
set.seed(1)

# Create three vectors
x <- runif(10)
y <- runif(10)
z <- runif(10)

# Create origin_fn that gets the median of each dimension
median_origin_fn <- create_origin_fn(median)

# Use median_origin_fn
median_origin_fn(x, y, z)

# Should be the same as
c(median(x), median(y), median(z))

}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
