% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_distances.R
\name{expand_distances}
\alias{expand_distances}
\title{Expand the distances to an origin}
\usage{
expand_distances(
  data,
  cols = NULL,
  multiplier = NULL,
  multiplier_fn = NULL,
  origin = NULL,
  origin_fn = NULL,
  exponentiate = FALSE,
  add_one_exp = TRUE,
  suffix = "_expanded",
  keep_original = TRUE,
  mult_col_name = ifelse(isTRUE(exponentiate), ".exponent", ".multiplier"),
  origin_col_name = ".origin",
  overwrite = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{vector}.}

\item{cols}{Names of columns in \code{`data`} to expand coordinates of.
Each column is considered a dimension.}

\item{multiplier}{Constant to multiply/exponentiate the distances to the origin by.

\strong{N.B.} When \code{`exponentiate`} is \code{TRUE}, the \code{`multiplier`} becomes an \emph{exponent}.}

\item{multiplier_fn}{Function for finding the \code{`multiplier`}.

\strong{Input}: Each column will be passed as a \code{vector} in the order of \code{`cols`}.

\strong{Output}: A \code{numeric scalar}.}

\item{origin}{Coordinates of the origin to expand around.
A scalar to use in all dimensions
or a \code{vector} with one scalar per dimension.

\strong{N.B.} Ignored when \code{`origin_fn`} is not \code{NULL}.}

\item{origin_fn}{Function for finding the origin coordinates.

\strong{Input}: Each column will be passed as a \code{vector} in the order of \code{`cols`}.

\strong{Output}: A \code{vector} with one scalar per dimension.

Can be created with \code{\link[rearrr:create_origin_fn]{create_origin_fn()}} if you want to apply
the same function to each dimension.

E.g. \code{`create_origin_fn(median)`} would find the median of each column.

\strong{Built-in functions} are \code{\link[rearrr:centroid]{centroid()}},
\code{\link[rearrr:most_centered]{most_centered()}},
and \code{\link[rearrr:midrange]{midrange()}}}

\item{exponentiate}{Whether to exponentiate instead of multiplying. (Logical)}

\item{add_one_exp}{Whether to add \code{1} to the distances
before exponentiating to ensure they don't contract when between \code{0} and \code{1}.
The added value is subtracted after the exponentiation. (Logical)

The distances to the origin (\code{`d`}) are exponentiated as such:

\code{d <- d + 1}

\code{d <- d ^ multiplier}

\code{d <- d - 1}

\strong{N.B.} Ignored when \code{`exponentiate`} is \code{FALSE}.}

\item{suffix}{Suffix to add to the names of the generated columns.

Use an empty string (i.e. \code{""}) to overwrite the original columns.}

\item{keep_original}{Whether to keep the original columns. (Logical)

Some columns may have been overwritten, in which case only the newest versions are returned.}

\item{mult_col_name}{Name of new column with the \code{`multiplier`}.
If \code{NULL}, no column is added.}

\item{origin_col_name}{Name of new column with the origin coordinates.
If \code{NULL}, no column is added.}

\item{overwrite}{Whether to allow overwriting of existing columns. (Logical)}
}
\value{
\code{data.frame} (\code{tibble}) with the expanded columns,
along with the applied multiplier/exponent and origin coordinates.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Moves the data points in n-dimensional space such that their distance
to a specified origin is increased/decreased.
A \code{`multiplier`} greater than 1 leads to expansion,
while a positive \code{`multiplier`} lower than 1 leads to contraction.

The origin can be supplied as coordinates or as a function that returns coordinates. The
latter can be useful when supplying a grouped \code{data.frame} and expanding around e.g. the centroid
of each group.

The multiplier/exponent can be supplied as a constant or as a function that returns a constant.
The latter can be useful when supplying a grouped \code{data.frame} and the multiplier/exponent depends
on the data in the groups.

For expansion in each dimension separately, use \code{\link[rearrr:expand_distances_each]{expand_distances_each()}}.

\strong{NOTE}: When exponentiating, the default is to first add \code{1} to the distances,
to ensure expansion even when the distance is between \code{0} and \code{1}.
If you need the purely exponentiated distances,
disable \code{`add_one_exp`}.
}
\details{
Increases the distance to the origin in n-dimensional space by
multiplying or exponentiating it by the multiplier.

We first move the origin to the zero-coordinates (e.g. \code{c(0, 0, 0)})
and normalize each vector to unit length. We then multiply this unit vector by the
multiplied/exponentiated distance and moves the origin back to its original coordinates.

The distance to the specified origin is calculated with:
\deqn{d(P1, P2) = sqrt( (x2 - x1)^2 + (y2 - y1)^2 + (z2 - z1)^2 + ... )}

Note: By default (when \code{`add_one_exp`} is \code{TRUE}),
we add \code{1} to the distance before the exponentiation
and subtract it afterwards. See \code{`add_one_exp`}.
}
\examples{
# Attach packages
library(rearrr)
library(dplyr)
library(purrr)
library(ggplot2)

# Set seed
set.seed(1)

# Create a data frame
df <- data.frame(
  "x" = runif(20),
  "y" = runif(20),
  "g" = rep(1:4, each = 5)
)

# Expand distances in the two dimensions (x and y)
# With the origin at x=0.5, y=0.5
# We multiply the distances by 2
expand_distances(
  data = df,
  cols = c("x", "y"),
  multiplier = 2,
  origin = c(0.5, 0.5)
)

# Expand distances in the two dimensions (x and y)
# With the origin at x=0.5, y=0.5
# We exponentiate the distances by 2
expand_distances(
  data = df,
  cols = c("x", "y"),
  multiplier = 2,
  exponentiate = TRUE,
  origin = 0.5
)

# Expand values in one dimension (x)
# With the origin at x=0.5
# We exponentiate the distances by 3
expand_distances(
  data = df,
  cols = c("x"),
  multiplier = 3,
  exponentiate = TRUE,
  origin = 0.5
)

# Expand x and y around the centroid
# We use exponentiation for a more drastic effect
# The add_one_exp makes sure it expands
# even when x or y is in the range [0, <1]
# To compare multiple exponents, we wrap the
# call in purrr::map_dfr
df_expanded <- purrr::map_dfr(
  .x = c(1, 3, 5),
  .f = function(exponent) {
    expand_distances(
      data = df,
      cols = c("x", "y"),
      multiplier = exponent,
      origin_fn = centroid,
      exponentiate = TRUE,
      add_one_exp = TRUE
    )
  }
)
df_expanded

# Plot the expansions of x and y around the overall centroid
ggplot(df_expanded, aes(x = x_expanded, y = y_expanded, color = factor(.exponent))) +
  geom_vline(
    xintercept = df_expanded[[".origin"]][[1]][[1]],
    size = 0.2, alpha = .4, linetype = "dashed"
  ) +
  geom_hline(
    yintercept = df_expanded[[".origin"]][[1]][[2]],
    size = 0.2, alpha = .4, linetype = "dashed"
  ) +
  geom_path(size = 0.2) +
  geom_point() +
  theme_minimal() +
  labs(x = "x", y = "y", color = "Exponent")

# Expand x and y around the centroid using multiplication
# To compare multiple multipliers, we wrap the
# call in purrr::map_dfr
df_expanded <- purrr::map_dfr(
  .x = c(1, 3, 5),
  .f = function(multiplier) {
    expand_distances(df,
      cols = c("x", "y"),
      multiplier = multiplier,
      origin_fn = centroid,
      exponentiate = FALSE
    )
  }
)
df_expanded

# Plot the expansions of x and y around the overall centroid
ggplot(df_expanded, aes(x = x_expanded, y = y_expanded, color = factor(.multiplier))) +
  geom_vline(
    xintercept = df_expanded[[".origin"]][[1]][[1]],
    size = 0.2, alpha = .4, linetype = "dashed"
  ) +
  geom_hline(
    yintercept = df_expanded[[".origin"]][[1]][[2]],
    size = 0.2, alpha = .4, linetype = "dashed"
  ) +
  geom_path(size = 0.2, alpha = .8) +
  geom_point() +
  theme_minimal() +
  labs(x = "x", y = "y", color = "Multiplier")

#
# Contraction
#

# Group-wise contraction to create clusters
df_contracted <- df \%>\%
  dplyr::group_by(g) \%>\%
  expand_distances(
    cols = c("x", "y"),
    multiplier = 0.07,
    suffix = "_contracted",
    origin_fn = centroid
  )

# Plot the clustered data point on top of the original data points
ggplot(df_contracted, aes(x = x_contracted, y = y_contracted, color = factor(g))) +
  geom_point(aes(x = x, y = y, color = factor(g)), alpha = 0.3, shape = 16) +
  geom_point() +
  theme_minimal() +
  labs(x = "x", y = "y", color = "g")
}
\seealso{
Other mutate functions: 
\code{\link{apply_transformation_matrix}()},
\code{\link{cluster_groups}()},
\code{\link{dim_values}()},
\code{\link{expand_distances_each}()},
\code{\link{flip_values}()},
\code{\link{roll_values}()},
\code{\link{rotate_2d}()},
\code{\link{rotate_3d}()},
\code{\link{shear_2d}()},
\code{\link{shear_3d}()},
\code{\link{swirl_2d}()},
\code{\link{swirl_3d}()}

Other expander functions: 
\code{\link{expand_distances_each}()}

Other distance functions: 
\code{\link{closest_to}()},
\code{\link{dim_values}()},
\code{\link{distance}()},
\code{\link{expand_distances_each}()},
\code{\link{furthest_from}()},
\code{\link{swirl_2d}()},
\code{\link{swirl_3d}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{distance functions}
\concept{expander functions}
\concept{mutate functions}
