% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{positioning_rearranger}
\alias{positioning_rearranger}
\title{Wrapper for running positioning rearrange methods}
\usage{
positioning_rearranger(
  data,
  col = NULL,
  position = NULL,
  shuffle_sides = FALSE,
  what = "max"
)
}
\arguments{
\item{data}{\code{data frame} or \code{vector}.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{data} is a \code{data frame},
the row numbers are used.}

\item{position}{Index or quantile (in \code{0-1}) at which to position the element of interest.}

\item{shuffle_sides}{Whether to shuffle which elements are left and right of the position. (Logical)}

\item{what}{What to position. "max" or "min". (Character)}
}
\value{
Sorted \code{data frame} / \code{vector}.
}
\description{
Wrapper for running positioning rearrange methods
}
\keyword{internal}
