% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{rearranger_}
\alias{rearranger_}
\title{Wrapper for running rearranging methods}
\usage{
rearranger_(
  data,
  rearrange_fn,
  check_fn,
  cols = NULL,
  allowed_types = c("numeric", "factor", "character"),
  col = deprecated(),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} or \code{vector}.}

\item{rearrange_fn}{Rearrange function to apply.}

\item{check_fn}{Function with checks post-preparation of \code{`data`} and \code{`cols`}.
Should not return anything.}

\item{cols}{Column(s) to create sorting factor by. When \code{NULL} and \code{`data`} is a \code{data.frame},
the row numbers are used.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{`data`} is a \code{data.frame},
the row numbers are used.}

\item{...}{Named arguments for the \code{`rearrange_fn`}.}
}
\value{
The sorted \code{data.frame} (\code{tibble}) / \code{vector}.
Optionally with (a) sorting factor(s) added.

When \code{`data`} is a \code{vector} and
no extra factors are returned by \code{`rearrange_fn`},
the output will be a \code{vector}. Otherwise, a \code{data.frame}.
}
\description{
Wrapper for running rearranging methods
}
\keyword{internal}
