% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_windows.R
\name{rev_windows}
\alias{rev_windows}
\title{Reverse order window-wise}
\usage{
rev_windows(data, window_size, keep_windows = FALSE, factor_name = ".window")
}
\arguments{
\item{data}{\code{data frame} or \code{vector}.}

\item{window_size}{Size of the windows.}

\item{keep_windows}{Whether to keep the factor with window identifiers. (Logical)}

\item{factor_name}{Name of windows factor.

N.B. Only used when \code{keep_windows} is \code{TRUE}.}
}
\value{
The sorted \code{data frame} / \code{vector}.
Optionally with the windows factor added.

When \code{data} is a \code{vector} and \code{keep_windows} is \code{FALSE},
the output will be a \code{vector}. Otherwise, a \code{data frame}.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

The values are windowed and reversed within windows.

\strong{Example}:

The column values:

\code{c(1, 2, 3, 4, 5, 6)}

With \code{window_size = 3}

Are \strong{ordered as}:

\code{c(3, 2, 1, 6, 4, 5)}
}
\examples{
\donttest{
# Attach packages
library(rearrr)

# Set seed
set.seed(1)

# Create a data frame
df <- data.frame(
  "index" = 1:10,
  "A" = sample(1:10),
  "B" = runif(10),
  "C" = LETTERS[1:10],
  "G" = c(1, 1, 1, 1, 1,
          2, 2, 2, 2, 2),
  stringsAsFactors = FALSE
)

# For vector
rev_windows(1:10, window_size = 3)

# For data frame
rev_windows(df, window_size = 3)
rev_windows(df, window_size = 3, keep_windows = TRUE)

# Grouped by G
df \%>\%
  dplyr::select(G, index) \%>\%  # For clarity
  dplyr::group_by(G) \%>\%
  rev_windows(window_size = 3)

# Plot the extreme pairs
plot(x = 1:10,
     y = rev_windows(1:10, window_size = 3))
}
}
\seealso{
Other rearrange functions: 
\code{\link{center_max}()},
\code{\link{center_min}()},
\code{\link{pair_extremes}()},
\code{\link{position_max}()},
\code{\link{position_min}()},
\code{\link{rearrange}()}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\concept{rearrange functions}
