% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{rev_windows_rearranger}
\alias{rev_windows_rearranger}
\title{Wrapper for running centering rearrange methods}
\usage{
rev_windows_rearranger(
  data,
  window_size,
  keep_windows = FALSE,
  factor_name = ".window"
)
}
\arguments{
\item{data}{\code{data frame} or \code{vector}.}

\item{window_size}{Size of the windows. (Logical)}

\item{keep_windows}{Whether to keep the factor with window identifiers. (Logical)}

\item{factor_name}{Name of the factor with window identifiers. (Logical)

N.B. Only used when \code{keep_windows} is \code{TRUE}.}
}
\value{
The sorted \code{data frame} / \code{vector}.
Optionally with the windows factor added.

When \code{data} is a \code{vector} and \code{keep_windows} is \code{FALSE},
the output will be a \code{vector}. Otherwise, a \code{data frame}.
}
\description{
Wrapper for running centering rearrange methods
}
\keyword{internal}
