% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{rev_windows_rearranger_}
\alias{rev_windows_rearranger_}
\title{Wrapper for running centering rearrange methods}
\usage{
rev_windows_rearranger_(
  data,
  window_size,
  factor_name = ".window",
  overwrite = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{vector}.}

\item{window_size}{Size of the windows. (Logical)}

\item{factor_name}{Name of the factor with window identifiers.
If \code{`NULL`}, no column is added.}

\item{overwrite}{Whether to allow overwriting of existing columns. (Logical)}
}
\value{
The sorted \code{data.frame} (\code{tibble}) / \code{vector}.
Optionally with the windows factor added.

When \code{`data`} is a \code{vector} and \code{`keep_windows`} is \code{`FALSE`},
the output will be a \code{vector}. Otherwise, a \code{data.frame}.
}
\description{
Wrapper for running centering rearrange methods
}
\keyword{internal}
