% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangers.R
\name{by_distance_rearranger}
\alias{by_distance_rearranger}
\title{Wrapper for running closest to / furthest from rearrange methods}
\usage{
by_distance_rearranger(
  data,
  col,
  target = NULL,
  target_fn = NULL,
  shuffle_ties = FALSE,
  decreasing = FALSE
)
}
\arguments{
\item{data}{\code{data frame} or \code{vector}.}

\item{col}{Column to create sorting factor by. When \code{NULL} and \code{data} is a \code{data frame},
the row numbers are used.}

\item{target}{Target value. (Logical)}

\item{target_fn}{Function for extracting target value. (Logical)

\strong{N.B.} Either \code{target} or \code{target_fn} should be specified.

\strong{N.B.} When \code{data} is grouped,
the \code{target_fn} function is applied group-wise.}

\item{shuffle_ties}{Whether to shuffle elements with the same distance to the target. (Logical)}

\item{decreasing}{Whether to prder by decreasing distances to the target. (Logical)}
}
\value{
The sorted \code{data frame} / \code{vector}.
}
\description{
Wrapper for running closest to / furthest from rearrange methods
}
\keyword{internal}
