% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_centered.R
\name{most_centered}
\alias{most_centered}
\title{Find the coordinates for data point closest to the centroid}
\usage{
most_centered(...)
}
\arguments{
\item{...}{Numeric vectors.}
}
\value{
\code{vector} with the coordinates for the data point closest to the centroid.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Returns the coordinates for the
data point with the shortest distance to the centroid.
}
\examples{
\donttest{
# Attach packages
library(rearrr)

# Set seed
set.seed(1)

# Create three vectors
x <- runif(10)
y <- runif(10)
z <- runif(10)

# Find coordinates of data point
# closest to the centroid
most_centered(x, y, z)

# Compare to centroid coordinates
centroid(x, y, z)
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
