% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutators.R
\name{multi_mutator}
\alias{multi_mutator}
\title{Wrapper for running multicolumn mutator methods}
\usage{
multi_mutator(
  data,
  mutate_fn,
  check_fn,
  cols = NULL,
  suffix = "_mutated",
  force_df = TRUE,
  allowed_types = c("numeric", "factor"),
  min_dims = 1,
  keep_original = TRUE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} or \code{vector}.}

\item{mutate_fn}{Mutator to apply.}

\item{check_fn}{Function with checks post-preparation of \code{`data`} and \code{`cols`}.
Should not return anything.}

\item{cols}{Columns to mutate values of. Must be specified when \code{`data`} is a \code{data.frame}.}

\item{suffix}{Suffix to add to the names of the generated columns.

Use an empty string (i.e. \code{""}) to overwrite the original columns.}

\item{force_df}{Whether to return a \code{data.frame} when \code{`data`} is a \code{vector}.}

\item{allowed_types}{Allowed types of the \code{`col(s)`} colums. The type restrictions do not apply to
columns not mentioned in the \code{`col(s)`} argument.}

\item{min_dims}{Minimum number of dimensions (cols) after preparations. When \code{`data`} is a \code{vector}
setting \code{`min_dims`} to \code{2} will use both the index and the values as columns.}

\item{keep_original}{Whether to keep the original columns. (Logical)

Some columns may have been overwritten, in which case only the newest versions are returned.}

\item{...}{Named arguments for the \code{`mutate_fn`}.}
}
\value{
The mutated \code{data.frame} (\code{tibble}).
}
\description{
Wrapper for running multicolumn mutator methods
}
\keyword{internal}
